Grind main/monobj: single-precision float compares in checkCol cone-hit guards (#16526)
Drop spurious static_cast<double> wrapping on the kMonObjZero>=targetDist
and (sideDist-radius)>targetDist cone-detection guards so MWCC emits
single-precision fcmpo against the float operands (matching the target's
lfs/fsubs shape) instead of promoting to double (lfd/fsub+frsp).
checkCol 80.50 -> 80.73; unit 90.42235 -> 90.4373.

+++ src/monobj.cpp
@@ -2212,25 +2212,22 @@ void CGMonObj::checkCol(int flags, float rotY, float distance, float* hitScale, 
 				targetDelta.z = delta.z;
 			}
 			float targetDist = PSVECMag(&targetDelta);
-			if (static_cast<double>(kMonObjZero) >= static_cast<double>(targetDist)) {
+			if (kMonObjZero >= targetDist) {
 				continue;
 			}
 
 			CVector targetDir;
 			PSVECNormalize(&targetDelta, reinterpret_cast<Vec*>(&targetDir));
 			float dot = PSVECDotProduct(reinterpret_cast<Vec*>(&forward), reinterpret_cast<Vec*>(&targetDir));
-			if ((static_cast<double>(
-			         static_cast<float>(static_cast<double>(sideDist) -
-			                            static_cast<double>(object->m_bodyEllipsoidRadius))) >
-			     static_cast<double>(targetDist)) ||
+			if (((sideDist - object->m_bodyEllipsoidRadius) > targetDist) ||
 				((static_cast<double>(kMonObjZero) != static_cast<double>(halfAngle)) &&
-				 (static_cast<double>(kMonObjZero) >= static_cast<double>(dot)))) {
+				 (kMonObjZero >= dot))) {
 				continue;
 			}
 
 			float angle = static_cast<float>(acos(static_cast<double>(dot)));
 			if ((static_cast<double>(kMonObjZero) != static_cast<double>(halfAngle)) &&
-				(static_cast<double>(angle) >= static_cast<double>(halfAngle))) {
+				(angle >= halfAngle)) {
 				continue;
 			}
